Field layout in class Arguments is incompatible w\ 64-bit archs.
The length_ field must be defined as intptr_t rather than int.  This
is due to the fact that we place native argc/argv values into stack
slots (via SetFrameSlot) and then interpret the slots as a an instance
of Arguments class.

Little endian architectures get 'lucky' that the layout happens to
work with implicit padding.  Big endian is not as lucky.

See Runtime_ArrayConstructor for an example.