PPC: Manipulate external arrays according to user-specified endianness mode.
The following flags define the endianness mode for memory accesses to
external array elements:

  --little-endian-external-arrays
  --big-endian-external-arrays

The flags are mutually exclusive.

The default behavior (where neither flag is set) is to conform to the
endianness mode of the host architecture.

These flags are intended to help work around issues caused by
javascript code with endianness assumptions which do not match that of
the host architecture.