Revert changes to ool constant pool header layout.
Rather than increasing the capacity of the ConstantPoolArray's small
section to match the longer (16-bit) reach of PPC's load instruction
(at the expense of a larger header to describe the layout), the PPC
implementation now utilizes the extended section to satisfy that
reach.  I.e. all entries (regardless of their section) are reachable
with a single load instruction.

As before, this implementation does not support an unlimited constant
pool size (which would require a multi-instruction sequence).  [See
ARM commit e27ab337 for a reference on the changes required to support
the longer instruction sequence.]  Note, however, that going down that
path will necessarily generate that longer sequence for all extended
section accesses since the placement of a given entry within the
section is not known at the time of code generation.