Address failures in test/mjsunit/asm/uint32div.js in ppc.debug mode
In debug mode on ppc and ppc64 there were failures in the test
test/mjsunit/asm/uint32div.js.  In the failing cases the mesage was

    Failure: expected <0> found <-1>

The method was being compiled with turbofan and in the optimized code
divwu was being used to do the division.  In the emulator we returned
-1 in the case of a division by 0 as the spec says the result is
undefined.  The problem was that we were returning kUint32DivIsSafe
to turbofan. We were just getting "lucky" in release mode because
the real instruction happened to return 0 even though this is not
guarranteed to be the case.

As part of debugging this issue we also noticed the disassembler
was missing the definition for divwu so this patch adds that as well.

	modified:   src/compiler/ppc/instruction-selector-ppc.cc
	modified:   src/ppc/disasm-ppc.cc