Stack sampling test failures in ppc and ppc64 simulation mode
The StackDepthIsConsistent and StackFramesConsistent tests were failing
in ppc and ppc64 simulation mode.  The cause turned out to be that
we were missing the ppc and ppc64 cases in test/cctest/test-sampler-api.cc
were there is platform specific code to set the registers when running in
simulated mode.  Since this code was missing the valid stack range was
incorrectly including 0 at the bottom end which caused problems because
a 0 caller fp value exists in the outermost frame.

	modified:   test/cctest/test-sampler-api.cc