AIX: Temporarily disable ceil/trunc/round tests in debug mode.
This is related to non-standard behavior in AIX's implementation for
ceil/trunc/round routines which return 0.0 instead of -0.0 when
applicable.  This was previously worked around by passing
-mcpu=power5+ -mfprnd to gcc in order to replace those calls with
single-instruction implementations (frip/friz/frin) which have the
correct behavior.  Unfortunately, this optimization is disabled at
with -O0 -- and thus the failures of these tests in debug mode only.

Using -Og instead of -O0 should resolve the problem, however this
option is only available in GCC 4.8.

Deferring this until that version is available and installed on the
AIX build machines.