Ensure Deoptimizer input frame registers are initialized to Smi values.
This fixes the following 64-bit debug-mode testcase failure:

  cctest --verify-heap test-debug/DebugBreakLoop

During deoptimization -- converting an optimized from to an
unoptimized one -- it attempts to retrieve the context pointer from r3
in the intput frame (based on the deopt data).  r3's value was 0xc --
a HeapObject under V8_PPC_TAGGING_OPT.  With the --verify-heap option
specified in the debug run, we then walk the heap until we visit the
DeoptimizedFrameInfo object and fault attempting to visit the context
pointer.