Introduce alternate OOL constant pool mode.
This provides an alternate mode for out-of-line constant pools.  The
new mode supports the placement of a contiguous constant pool at the
end of the code buffer rather than a separate heap object.  It
leverages the traditional mode's common code infrastructure for
maintaining a dedicated constant pool base register, but does not
require the ConstantPoolArray heap object.