Move to 4.8.3 for compilation on AIX - test-run-machops.cc
An earlier commit for the move to the 4.8.3 compiler removed
the exclude for tests in test-run-machops.cc.  To make these
pass it also required that we enable -Og instead of -O0 for
AIX in order to work around a compiler bug.  This commit
adds that change.  Doing this also made the compiler complain
about values which could have been used without initialization
so these values in src/heap/heap.cc were set to NULL.  Will
need to revisit why this is necessary only with -Og as part
of contribution back to google repos

	modified:   build/standalone.gypi
	modified:   src/heap/heap.cc