PPC: [turbofan] Materialize JSFunction from frame if possible.
Port 725cdc533cb4e31c042d32ce1979012b5bd99ced

Original commit message:
This reduces the overhead of recursive calls when context specialization
is enabled. Based on this it might be possible to further reduce the
overhead by also specializing the call itself.

As a drive-by-fix, port the fast context materialization optimization to
arm and arm64, that was previously only supported on x64 and ia32.