PPC: Resolve references to "this" the same way as normal variables
Port 18619d355192e2699203d12d9ebb9caea107b693

Original commit message:
Make the parser handle references to "this" as unresolved variables, so the
same logic as for the rest of function parameters is used for the receiver.
Minor additions to the code generation handle copying the receiver to the
context, along with the rest of the function parameters.

Based on work by Adrian Perez de Castro <aperez@igalia.com>.

R=wingo@igalia.com, dstence@us.ibm.com,  michael_dawson@ca.ibm.com